+++ Src/Obfy.Core/Utilities/DecryptorIl.cs
@@ -1,6 +1,5 @@
 using dnlib.DotNet;
 using dnlib.DotNet.Emit;
-using Obfy.Core.Models;
 
 namespace Obfy.Core.Utilities;
 
@@ -209,10 +208,9 @@ internal static class DecryptorIl
         FieldDef cacheField,
         FieldDef indexXorField,
         MethodDef bytesDecrypt,
-        EncryptionAlgorithm algorithm,
         string name = "Decrypt")
     {
-        _ = algorithm;
+        ArgumentException.ThrowIfNullOrWhiteSpace(name);
         var method = new MethodDefUser(
             name,
             MethodSig.CreateStatic(module.CorLibTypes.String, module.CorLibTypes.Int32),